Feat : 예외처리 수정

+++ src/main/java/com/ff1451/movie_review/service/ReviewService.java
@@ -6,12 +6,15 @@ import com.ff1451.movie_review.dto.review.ReviewUpdateRequest;
 import com.ff1451.movie_review.entity.Movie;
 import com.ff1451.movie_review.entity.Review;
 import com.ff1451.movie_review.entity.User;
+import com.ff1451.movie_review.exception.CustomException;
+import com.ff1451.movie_review.exception.ErrorCode;
 import com.ff1451.movie_review.repository.MovieRepository;
 import com.ff1451.movie_review.repository.ReviewRepository;
 import com.ff1451.movie_review.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 
 import java.util.List;
 
@@ -32,7 +35,7 @@ public class ReviewService {
 
     public ReviewResponse getReviewById(Long id) {
         Review review = reviewRepository.findById(id)
-            .orElseThrow(()-> new RuntimeException("해당 리뷰를 찾을 수 없습니다."));
+            .orElseThrow(()-> new CustomException(ErrorCode.REVIEW_NOT_FOUND));
         return ReviewResponse.from(review);
     }
 
@@ -53,10 +56,15 @@ public class ReviewService {
     @Transactional
     public ReviewResponse createReview(ReviewRequest request) {
         Movie movie = movieRepository.findById(request.movieId())
-            .orElseThrow(()-> new RuntimeException("찾을 수 없음"));
+            .orElseThrow(()-> new CustomException(ErrorCode.MOVIE_NOT_FOUND));
 
         User user = userRepository.findById(request.userId())
-            .orElseThrow(()-> new RuntimeException("찾을 수 없음"));
+            .orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
+
+        reviewRepository.findByMovieIdAndUserId(request.movieId(),request.userId())
+            .ifPresent(review -> {
+                throw new CustomException(ErrorCode.REVIEW_EXIST);
+            });
 
         Review review = new Review(
             movie,
@@ -71,9 +79,13 @@ public class ReviewService {
     }
 
     @Transactional
-    public ReviewResponse updateReview(Long id, ReviewUpdateRequest request) {
+    public ReviewResponse updateReview(Long id, Long userId, ReviewUpdateRequest request) {
         Review review = reviewRepository.findById(id)
-            .orElseThrow(()->new RuntimeException("리뷰를 찾을 수 없음"));
+            .orElseThrow(()->new CustomException(ErrorCode.REVIEW_NOT_FOUND));
+
+        if (!review.getUser().getId().equals(userId)) {
+            throw new CustomException(ErrorCode.UNAUTHORIZED_ACTION);
+        }
 
         request.rating().ifPresent(review::setRating);
         request.reviewText().ifPresent(review::setReviewText);
@@ -85,9 +97,14 @@ public class ReviewService {
     }
 
     @Transactional
-    public void deleteReview(Long id) {
+    public void deleteReview(Long id, Long userId) {
         Review review = reviewRepository.findById(id)
-            .orElseThrow(()->new RuntimeException("리뷰를 찾을 수 없습니다."));
+            .orElseThrow(()->new CustomException(ErrorCode.REVIEW_NOT_FOUND));
+
+        if (!review.getUser().getId().equals(userId)) {
+            throw new CustomException(ErrorCode.UNAUTHORIZED_ACTION);
+        }
+
         reviewRepository.deleteById(id);
         updateMovieRating(review.getMovie().getId());
     }
